+++ NEWS.md
@@ -1,4 +1,5 @@
 ## 0.4.3.dev0 to 0.5.0 (unreleased so far)
+ * Contacts: Adds example script for importing a manual whitelist.
  * Technical: Module called "intelmq-fody-backend" now
  * Technical: Compatible with psycopg2 coming with Ubuntu 16.04LTS now.